0x10029FE0 BrMat4Perspective7 T4: explicit (float) cast on the w product pins the rounding store before the h push

+++ src/core/slice4_50.c
@@ -176,16 +176,17 @@ int BrMat4Perspective7(BrMat4 *pM, uint16_t *pPerspNorm,
      * multiplies first, pops after, and homes h in n's slot.  Naming -h
      * gives size- and instruction-exact output; naming -w as well is
      * inert, naming -w alone is worse (15).
-     * RESIDUE, 9 diff bytes, register-blind 0+0, ONE region at +0x3b: the
-     * original stores w to its slot and loads its bits into ecx BEFORE
-     * `push eax` (h); ours pushes h first.  Dead: every order of the
-     * nh/nw statements and of the declarations, `aspect * h`, `w` spelled
-     * inline in the call, `h` spelled inline (worse, 21), a `double` ty,
-     * `n`/`f`/`scale` copied into locals, `n * tan(...)` with no ty. */
+     * The last region (w's store/load ordered before `push eax`) fell to
+     * the explicit `(float)` cast on the w product, 2026-09-09: the cast
+     * pins the rounding store next to the multiply, so w's bits are in ecx
+     * before h is pushed.  (Dead first: every order of the nh/nw statements
+     * and declarations, `aspect * h`, inline spellings, a double ty,
+     * copied scalars, per-part casts on h/ty, 0.0f-h, arg tweaks, every
+     * slot in the TU.) */
     float ty, h, w, nh;
     ty = (float)tan((double)fovyDegrees * 0.0087266462599716477);
     h = n * ty;
-    w = h * aspect;
+    w = (float)(h * aspect);
     nh = -h;
     ((int (*)(BrMat4 *, float, float, float, float, float, float, float))BrMat4Frustum)
         (pM, -w, w, nh, h, n, f, scale);
